// typings.d.ts
export type Programa = {
  Id: int;
  Nombre: string;
  Sigla: string;
  cantidad: number;
  preguntas: any;
  IdSubSede: string;
  Periodicidad: string;
  Cu: string;
};

export type Administrativo = {
  Id: int;
  Nombre: string;
  Apellidos: string;
  TipoDocumento: string;
  Documento: int;
  Cargo: string;
  Correo: string;
  Perfil: string;
  Usuario: string;
  Pass: string;
  Imagen: string;
  Password: string;
  Genero: number | string;
  IndexDocumento: number;
  TipoUsuario: number;
  IndexRol: number;
  IdSubSede?: string;
  SubSede?: string;
};

export type Documento = {
  Id: int;
  Codigo: string;
  Nombre: string;
};

export type Jornada = {
  Id: int;
  Nombre: string;
  Sigla: string;
};

export type SemestreAcademico = { Id: number; Nombre: string; Numero: any };

export type Competencia = {
  Id: int;
  Nombre: string;
  Ponderacion: number;
  Orden: int;
  Abreviatura: string;
  TipoCompetencia: string;
  DocenteAsignado?: Docente;
  Hora?: string;
  Minutos?: string;
};

export type Docente = {
  Id: int;
  Nombre: string;
  Apellidos: string;
  TipoDocumento: string;
  Documento: int;
  Usuario: string;
  Password: string;
  Genero: string;
  DocumCodigo: string;
  Imagen: string;
  Correo: string;
  Pass: string;
  IdSubSede?: string;
  Programa?: string;
  Cu?: string;
};